Renaming lib from "client" to "frontend"
