+++ apps/cms/src/components/TogglePublishDocument.tsx
@@ -1,4 +1,4 @@
-import {TanamPublishStatus} from "@tanam/domain-client";
+import {TanamPublishStatus} from "@tanam/domain-frontend";
 import {Button, DatePicker, Modal} from "@tanam/ui-components";
 import {useParams} from "next/navigation";
 import {useMemo, useState} from "react";
