+++ apps/cms/src/hooks/useTanamDocumentTypes.tsx
@@ -1,4 +1,4 @@
-import {TanamDocumentTypeClient, UserNotification} from "@tanam/domain-client";
+import {TanamDocumentTypeClient, UserNotification} from "@tanam/domain-frontend";
 import {collection, doc, onSnapshot} from "firebase/firestore";
 import {useParams} from "next/navigation";
 import {useEffect, useState} from "react";
