--- libs/domain-client/README.md
@@ -1,7 +1,0 @@
-# domain-client
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test domain-client` to execute the unit tests via [Jest](https://jestjs.io).
