+++ libs/domain-frontend/.eslintrc.json
