+++ libs/domain-frontend/README.md
@@ -1,0 +1,7 @@
+# domain-frontend
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test domain-frontend` to execute the unit tests via [Jest](https://jestjs.io).
