+++ libs/domain-frontend/project.json
@@ -1,12 +1,12 @@
 {
-  "name": "domain-client",
+  "name": "domain-frontend",
   "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/domain-client/src",
+  "sourceRoot": "libs/domain-frontend/src",
   "projectType": "library",
   "tags": [
     "lib:shared",
     "lib:frontend"
   ],
-  "// targets": "to see all targets run: nx show project domain-client --web",
+  "// targets": "to see all targets run: nx show project domain-frontend --web",
   "targets": {}
 }
