+++ libs/domain-frontend/src/models/TanamDocumentTypeClient.ts
