+++ libs/domain-frontend/src/models/TanamSpeechRecognition.ts
