+++ libs/domain-frontend/tsconfig.json
