+++ libs/domain-frontend/tsconfig.lib.json
