+++ libs/domain-frontend/tsconfig.spec.json
